refactor(timely): extract flatten-memories helper for plan/apply reuse
